+++ concordium-node/src/p2p/maintenance.rs
@@ -739,11 +739,10 @@ pub fn spawn(
                 && Instant::now().duration_since(last_buckets_cleaned)
                     >= Duration::from_millis(node.config.bucket_cleanup_interval)
             {
-                let decrement_bucket = |a: u64| {
-                    node.stats.recent_peers.with_label_values(&[a.to_string().as_str()]).dec();
-                };
-                write_or_die!(node.buckets())
-                    .clean_buckets(node.config.timeout_bucket_entry_period, decrement_bucket);
+                write_or_die!(node.buckets()).clean_buckets(
+                    node.config.timeout_bucket_entry_period,
+                    &node.stats.peer_bucket_size,
+                );
                 last_buckets_cleaned = Instant::now();
             }
         }